# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rais Motors showroom management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SuccessMessage {
  success: boolean;
  message?: string;
}

export interface ApiError {
  error: string;
}

export interface LoginInput {
  /** @minLength 1 */
  username: string;
  /** @minLength 1 */
  password: string;
}

export interface AuthSession {
  id: number;
  username: string;
  fullName: string;
  role: string;
}

export interface BikeInput {
  /** @minLength 1 */
  name: string;
  brand: string;
  model: string;
  category: string;
  color?: string;
  engineCc?: number | null;
  purchasePrice?: number;
  salePrice: number;
  stock: number;
  lowStockThreshold?: number;
  sku: string;
  imageUrl?: string | null;
  notes?: string | null;
}

export type Bike = BikeInput & {
  id: number;
  createdAt: string;
  updatedAt: string;
};

export interface CustomerInput {
  /** @minLength 1 */
  fullName: string;
  phone: string;
  email?: string | null;
  address?: string | null;
  cnic?: string | null;
  notes?: string | null;
}

export type Customer = CustomerInput & {
  id: number;
  createdAt: string;
};

export type SaleWithDetailsPaymentStatus =
  (typeof SaleWithDetailsPaymentStatus)[keyof typeof SaleWithDetailsPaymentStatus];

export const SaleWithDetailsPaymentStatus = {
  paid: "paid",
  partial: "partial",
  unpaid: "unpaid",
} as const;

export interface SaleWithDetails {
  id: number;
  invoiceNo: string;
  customerId: number;
  bikeId: number;
  quantity: number;
  unitPrice: number;
  discount: number;
  totalAmount: number;
  amountPaid: number;
  amountDue: number;
  paymentMethod: string;
  paymentStatus: SaleWithDetailsPaymentStatus;
  notes?: string | null;
  createdAt: string;
  customer: Customer;
  bike: Bike;
}

export type CustomerWithSales = Customer & {
  sales: SaleWithDetails[];
  totalSpent: number;
  totalDue: number;
};

export interface SaleInput {
  customerId: number;
  bikeId: number;
  /** @minimum 1 */
  quantity: number;
  unitPrice: number;
  discount?: number;
  amountPaid: number;
  paymentMethod: string;
  notes?: string | null;
}

export interface PaymentInput {
  /** @exclusiveMinimum 0 */
  amount: number;
  paymentMethod: string;
  note?: string | null;
}

export interface DashboardSummary {
  totalRevenue: number;
  totalSales: number;
  totalBikes: number;
  totalStock: number;
  totalCustomers: number;
  pendingPayments: number;
  monthRevenue: number;
  monthSales: number;
  weekRevenue: number;
  lowStockCount: number;
}

export interface MonthlyRevenuePoint {
  month: string;
  revenue: number;
  sales: number;
}

export interface WeeklySalesPoint {
  date: string;
  sales: number;
  revenue: number;
}

export interface BestSeller {
  bikeId: number;
  name: string;
  brand: string;
  unitsSold: number;
  revenue: number;
}

export interface PeriodStats {
  period: string;
  label: string;
  revenue: number;
  total: number;
  due: number;
  sales: number;
  uniqueCustomers: number;
  averageSale: number;
}

export interface ProfitStats {
  period: string;
  label: string;
  revenue: number;
  cost: number;
  grossProfit: number;
  expenses: number;
  netProfit: number;
  sales: number;
  margin: number;
}

export interface Bank {
  id: number;
  name: string;
  accountNumber?: string | null;
  openingBalance: number;
  balance: number;
  notes?: string | null;
  createdAt: string;
}

export interface BankInput {
  name: string;
  accountNumber?: string | null;
  openingBalance?: number;
  notes?: string | null;
}

export interface BankTransaction {
  id: number;
  bankId: number;
  bankName?: string | null;
  type: string;
  amount: number;
  description: string;
  refType?: string | null;
  refId?: number | null;
  createdAt: string;
}

export type BankTransactionInputType =
  (typeof BankTransactionInputType)[keyof typeof BankTransactionInputType];

export const BankTransactionInputType = {
  credit: "credit",
  debit: "debit",
} as const;

export interface BankTransactionInput {
  type: BankTransactionInputType;
  amount: number;
  description?: string;
}

export interface Expense {
  id: number;
  expenseDate: string;
  category: string;
  description: string;
  paidTo?: string | null;
  amount: number;
  bankId?: number | null;
  bankName?: string | null;
  paymentMethod: string;
  notes?: string | null;
  createdAt: string;
}

export interface ExpenseInput {
  expenseDate?: string;
  category?: string;
  description: string;
  paidTo?: string | null;
  amount: number;
  bankId?: number | null;
  paymentMethod?: string;
  notes?: string | null;
}

export interface ExpenseSummary {
  total: number;
  month: number;
  year: number;
}

export interface PurchaseOrder {
  id: number;
  orderNo: string;
  supplierName: string;
  bikeId: number;
  bikeName?: string | null;
  quantity: number;
  unitCost: number;
  totalCost: number;
  status: string;
  bankId?: number | null;
  bankName?: string | null;
  orderDate: string;
  receivedDate?: string | null;
  notes?: string | null;
  createdAt: string;
}

export interface PurchaseOrderInput {
  supplierName: string;
  bikeId: number;
  quantity: number;
  unitCost: number;
  bankId?: number | null;
  orderDate?: string;
  status?: string;
  notes?: string | null;
}

export type ListBikesParams = {
  search?: string;
  status?: ListBikesStatus;
};

export type ListBikesStatus =
  (typeof ListBikesStatus)[keyof typeof ListBikesStatus];

export const ListBikesStatus = {
  in_stock: "in_stock",
  low_stock: "low_stock",
  out_of_stock: "out_of_stock",
} as const;

export type ListCustomersParams = {
  search?: string;
};

export type ListSalesParams = {
  search?: string;
  paymentStatus?: ListSalesPaymentStatus;
};

export type ListSalesPaymentStatus =
  (typeof ListSalesPaymentStatus)[keyof typeof ListSalesPaymentStatus];

export const ListSalesPaymentStatus = {
  paid: "paid",
  partial: "partial",
  unpaid: "unpaid",
} as const;

export type GetPeriodStatsParams = {
  period?: GetPeriodStatsPeriod;
};

export type GetPeriodStatsPeriod =
  (typeof GetPeriodStatsPeriod)[keyof typeof GetPeriodStatsPeriod];

export const GetPeriodStatsPeriod = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
  yearly: "yearly",
  all: "all",
} as const;

export type GetProfitParams = {
  period?: GetProfitPeriod;
};

export type GetProfitPeriod =
  (typeof GetProfitPeriod)[keyof typeof GetProfitPeriod];

export const GetProfitPeriod = {
  daily: "daily",
  weekly: "weekly",
  monthly: "monthly",
  yearly: "yearly",
  all: "all",
} as const;

export type ListExpensesParams = {
  period?: string;
};