# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Rais Motors showroom management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getCurrentUser
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /bikes:
    get:
      operationId: listBikes
      tags: [bikes]
      summary: List all bikes
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: status
          schema:
            type: string
            enum: [in_stock, low_stock, out_of_stock]
      responses:
        "200":
          description: List of bikes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bike"
    post:
      operationId: createBike
      tags: [bikes]
      summary: Create a new bike
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BikeInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bike"
  /bikes/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getBike
      tags: [bikes]
      responses:
        "200":
          description: Bike
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bike"
    patch:
      operationId: updateBike
      tags: [bikes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BikeInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bike"
    delete:
      operationId: deleteBike
      tags: [bikes]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      parameters:
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
  /customers/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      operationId: getCustomer
      tags: [customers]
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerWithSales"
    patch:
      operationId: updateCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
...
  /banks:
    get:
      operationId: listBanks
      tags: [banks]
      responses:
        "200":
          description: All banks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bank"
    post:
      operationId: createBank
      tags: [banks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bank"
  /banks/{id}:
    patch:
      operationId: updateBank
      tags: [banks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bank"
    delete:
      operationId: deleteBank
      tags: [banks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /banks/{id}/transactions:
    get:
      operationId: listBankTransactions
      tags: [banks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BankTransaction"
    post:
      operationId: createBankTransaction
      tags: [banks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BankTransactionInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BankTransaction"
...
components:
  schemas:
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ApiError:
      type: object
      properties:
        error:
          type: string
      required: [error]
    LoginInput:
      type: object
      properties:
        username:
          type: string
          minLength: 1
        password:
          type: string
          minLength: 1
      required: [username, password]
    AuthSession:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        role:
          type: string
      required: [id, username, fullName, role]
    BikeInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        brand:
          type: string
        model:
          type: string
        category:
          type: string
        color:
          type: string
        engineCc:
          type: integer
          nullable: true
        purchasePrice:
          type: number
        salePrice:
          type: number
        stock:
          type: integer
        lowStockThreshold:
          type: integer
        sku:
          type: string
        imageUrl:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, brand, model, category, salePrice, stock, sku]
    Bike:
      allOf:
        - $ref: "#/components/schemas/BikeInput"
        - type: object
          properties:
            id:
              type: integer
            createdAt:
              type: string
            updatedAt:
              type: string
          required: [id, createdAt, updatedAt]
    CustomerInput:
      type: object
      properties:
        fullName:
          type: string
          minLength: 1
        phone:
          type: string
        email:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        cnic:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [fullName, phone]
    Customer:
      allOf:
        - $ref: "#/components/schemas/CustomerInput"
        - type: object
          properties:
            id:
              type: integer
            createdAt:
              type: string
          required: [id, createdAt]
    CustomerWithSales:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          properties:
            sales:
              type: array
              items:
                $ref: "#/components/schemas/SaleWithDetails"
            totalSpent:
              type: number
            totalDue:
              type: number
          required: [sales, totalSpent, totalDue]
...
    SaleWithDetails:
      type: object
      properties:
        id:
          type: integer
        invoiceNo:
          type: string
        customerId:
          type: integer
        bikeId:
          type: integer
        quantity:
          type: integer
        unitPrice:
          type: number
        discount:
          type: number
        totalAmount:
          type: number
        amountPaid:
          type: number
        amountDue:
          type: number
        paymentMethod:
          type: string
        paymentStatus:
          type: string
          enum: [paid, partial, unpaid]
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        customer:
          $ref: "#/components/schemas/Customer"
        bike:
          $ref: "#/components/schemas/Bike"
      required:
        - id
        - invoiceNo
        - customerId
        - bikeId
        - quantity
        - unitPrice
        - discount
        - totalAmount
        - amountPaid
        - amountDue
        - paymentMethod
        - paymentStatus
        - createdAt
        - customer
        - bike
...
    Bank:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        accountNumber:
          type: string
          nullable: true
        openingBalance:
          type: number
        balance:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, openingBalance, balance, createdAt]
    BankInput:
      type: object
      properties:
        name:
          type: string
        accountNumber:
          type: string
          nullable: true
        openingBalance:
          type: number
        notes:
          type: string
          nullable: true
      required: [name]
    BankTransaction:
      type: object
      properties:
        id:
          type: integer
        bankId:
          type: integer
        bankName:
          type: string
          nullable: true
        type:
          type: string
        amount:
          type: number
        description:
          type: string
        refType:
          type: string
          nullable: true
        refId:
          type: integer
          nullable: true
        createdAt:
          type: string
      required: [id, bankId, type, amount, description, createdAt]
    BankTransactionInput:
      type: object
      properties:
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        description:
          type: string
      required: [type, amount]
...