# lib/api-zod/src/generated/types/authSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Rais Motors showroom management API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthSession {
  id: number;
  username: string;
  fullName: string;
  role: string;
}